Create the phenotype model for acute diseases

Create the phenotype model for acute diseases

createAcutePhenotypeModel(
  connectionDetails,
  cdmDatabaseSchema,
  databaseId = "TestDB",
  cohortDatabaseSchema,
  cohortTable,
  outDatabaseSchema,
  modelOutputFileName = "train",
  evaluationOutputFileName = "eval",
  phenotypeEvaluationFileName = "results",
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId,
  excludedCovariateConceptIds = c(),
  includedCovariateIds = c(),
  addDescendantsToExclude = TRUE,
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  startDays = 0,
  endDays = 7,
  visitLength = 0,
  gender = c(8507, 8532),
  startDate = "19000101",
  endDate = "21000101",
  removeSubjectsWithFutureDates = TRUE,
  cdmVersion = "5",
  outFolder = getwd(),
  excludeModelFromEvaluation = FALSE,
  savePlpData = FALSE,
  createModel = TRUE,
  createEvaluationCohort = TRUE,
  cohortDefinitionsToTest = NULL
)

Arguments

connectionDetails
connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

cdmDatabaseSchema
The name of the database schema that contains the OMOP CDM instance. Requires read permissions to this database. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.

databaseId
Short name for the database (default="TestDB")

cohortDatabaseSchema
The name of the database schema that is the location where the cohort data used to define the at risk cohort is available. Requires read permissions to this database.

cohortTable
The tablename that contains the at risk cohort. The expectation is cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

outDatabaseSchema
The name of a database schema where the user has write capability. A temporary cohort table will be created here.

modelOutputFileName
A string designation for the training model file

evaluationOutputFileName
A string designation for the evaluation cohort file

phenotypeEvaluationFileName
A string designation for the .csv file with the results for the phenotype alogrithm evaluation

xSpecCohortId
The number of the "extremely specific (xSpec)" cohort definition id in the cohort table (for noisy positives)

xSensCohortId
The number of the "extremely sensitive (xSens)" cohort definition id in the cohort table (used to exclude subjects from the base population)

prevalenceCohortId
The number of the cohort definition id to determine the disease prevalence, (default=xSensCohortId)

excludedCovariateConceptIds
A list of conceptIds to exclude from featureExtraction. These should include all concept_ids that were used to define the xSpec model (default=NULL)

includedCovariateIds
A list of covariate IDs that should be restricted to.

addDescendantsToExclude
Should descendants of excluded concepts also be excluded? (default=FALSE)

mainPopulationCohortId
The number of the cohort ID to be used as a base population for the model (default=NULL)

mainPopulationCohortIdStartDay
The number of days relative to the mainPopulationCohortId cohort start date to begin including visits (default=0)

mainPopulationCohortIdEndDay
The number of days relative to the mainPopulationCohortId cohort start date to end including visits (default=0)

baseSampleSize
The maximum number of subjects in the evaluation cohort (default=2M)

lowerAgeLimit
The lower age for subjects in the model (default=NULL)

upperAgeLimit
The upper age for subjects in the model (default=NULL)

startDays
The days to include prior to the cohort start date (default=0)

endDays
The days to include after the cohort start date (default=7)

visitLength
The minimum length of index visit for noisy negative comparison (default=3)

gender
The gender(s) to be included (default c(8507, 8532))

startDate
The starting date for including subjects in the model (default=NULL)

endDate
The ending date for including subjects in the model (default=NULL)

removeSubjectsWithFutureDates
Should dates be checked to remove future dates (default=TRUE)

cdmVersion
The CDM version of the database (default=5)

outFolder
The folder where the output files will be written (default=working directory)

excludeModelFromEvaluation
Should subjects used in the model be excluded from the evaluation cohort (default = TRUE)

savePlpData
Should large PLP data file be saved (default=FALSE)

createModel
Run the function to create the diagnostic predictive model (default=TRUE)

createEvaluationCohort
Run the function to create the evaluation cohort (default=TRUE)

cohortDefinitionsToTest
A dataframe with cohorts to evaluate. Leave blank to not test any cohort definitions (default=Null) with the format: atlasId - The cohort ID in ATLAS; atlasName - The full name of the cohort; cohortId - The cohort ID to use in the package. Usually the same as the cohort ID in ATLAS; name - A short name for the cohort, to use to create file names. Do not use special characters; washoutPeriod - The mininum required continuous observation time prior to index date for subjects within the cohort to test

Details

Function to run through the complete phenotype evaluation process from model building to developing an evaluation cohort to using the evaluation cohort to determine the performance characteristics of phenotype algorithms for acute health conditions.