Create the default covariate settings

Create the default covariate settings for chronic diseases

createDefaultChronicCovariateSettings(
  excludedCovariateConceptIds = c(),
  includedCovariateIds = c(),
  addDescendantsToExclude = FALSE,
  startDays = 0,
  endDays = 9999
)

Arguments

excludedCovariateConceptIds
A list of conceptIds to exclude from featureExtraction. These should include all concept_ids that were used to define the xSpec model.

includedCovariateIds
A list of covariate IDs that should be restricted to.

addDescendantsToExclude
Should descendants of excluded concepts also be excluded?

startDays
The days to include prior to the cohort start date.

endDays
The days to include after the cohort start date.

Details

Function to create the default covariate settings for chronic diseases