Export the PheValuator results to CSV files

Export the PheValuator results to CSV files

exportPheValuatorResults(
  outputFolder,
  exportFolder = file.path(outputFolder, "export"),
  pheValuatorAnalysisList,
  packageName = NULL,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  baseUrl = NULL,
  cohortSetReference = NULL,
  minCellCount = 5
)

Arguments

outputFolder
Name of the folder where all the outputs will be written to.

exportFolder
Name of the folder where all the exported files will be written to.

pheValuatorAnalysisList
A list of objects of type pheValuatorAnalysis as created using the createPheValuatorAnalysis function.

packageName
The name of the package containing the cohort definitions. Can be left NULL if baseUrl and cohortSetReference have been specified.

cohortToCreateFile
The location of the cohortToCreate file within the package. Is ignored if cohortSetReference is specified.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI". Can be left NULL if no cohort meta-data needs to be exported.

cohortSetReference
A data frame with four columns, as described in the details. Can be left NULL if packageName and cohortToCreateFile have been specified.

minCellCount
The minimum cell count for fields contains person counts or fractions.

Details

Currently three ways of executing this function are supported, which determines how the meta-data on the cohorts is retrieved:

Embedded in a study package, assuming the cohort definitions are stored in that package using the ROhdsiWebApi::insertCohortDefinitionSetInPackage. In this case, the packageName and cohortToCreateFile should be populated.

Using a WebApi interface to retrieve the cohort definitions. In this case, the baseUrl and cohortSetReference should be populated.

Skipping meta-data on the cohorts altogether. In this case, only the cohortSetReference needs to be populated.

The cohortSetReference argument must be a data frame with the following columns:

atlasId

The cohort ID in ATLAS.

atlasName

The full name of the cohort. This will be shown in the Shiny app.

cohortId

The cohort ID to use in the package. Usually the same as the cohort ID in ATLAS.

name

A short name for the cohort, to use to create file names. do not use special characters.