Create a parameter object for the function createEvaluationCohort

Create a parameter object for the function createEvaluationCohort

createCreateEvaluationCohortArgs(
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId,
  xSpecCohortSize = 5000,
  covariateSettings = createDefaultChronicCovariateSettings(excludedCovariateConceptIds
    = c(), addDescendantsToExclude = TRUE),
  modelPopulationCohortId = 0,
  modelPopulationCohortIdStartDay = 0,
  modelPopulationCohortIdEndDay = 0,
  evaluationPopulationCohortId = 0,
  evaluationPopulationCohortIdStartDay = 0,
  evaluationPopulationCohortIdEndDay = 0,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 0,
  visitType = c(9201, 9202, 9203, 581477, 262),
  gender = c(8507, 8532),
  race = 0,
  ethnicity = 0,
  startDate = "19001010",
  endDate = "21000101",
  modelId = "main",
  evaluationCohortId = "main",
  excludeModelFromEvaluation = FALSE,
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE,
  modelType = "acute"
)

Arguments

xSpecCohortId
The number of the "extremely specific (xSpec)" cohortdefinition id in the cohort table (for noisy positives).

xSensCohortId
The number of the "extremely sensitive (xSens)" cohortdefinition id in the cohort table (for noisy negatives).

prevalenceCohortId
The number of the cohort definition id to determine the disease prevalence.

xSpecCohortSize
The recommended xSpec sample size to use in model (default = 5000)

covariateSettings
A covariateSettings object as generated using createCovariateSettings().

modelPopulationCohortId
The number of the cohort to be used as a base population for the model. If set to 0, the entire database population will be used.

modelPopulationCohortIdStartDay
The number of days relative to the mainPopulationCohortId cohort start date to begin including visits.

modelPopulationCohortIdEndDay
The number of days relative to the mainPopulationCohortId cohort start date to end including visits.

evaluationPopulationCohortId
The number of the cohort to be used as a base population for the evalution cohort. If set to 0, the entire database population will be used.

evaluationPopulationCohortIdStartDay
The number of days relative to the evaluationPopulationCohortId cohort start date to begin including visits.

evaluationPopulationCohortIdEndDay
The number of days relative to the evaluationPopulationCohortId cohort start date to end including visits.

baseSampleSize
The maximum number of subjects in the evaluation cohort.

lowerAgeLimit
The lower age for subjects in the model.

upperAgeLimit
The upper age for subjects in the model.

visitLength
The minimum length of index visit for acute outcomes.

visitType
The concept_id for the visit type.

gender
The gender(s) to be included.

race
The race(s) to be included.

ethnicity
The ethnicity(s) to be included.

startDate
The starting date for including subjects in the model.

endDate
The ending date for including subjects in the model.

modelId
A string used to generate the file names for this model.

evaluationCohortId
A string used to generate the file names for this evaluation cohort.

excludeModelFromEvaluation
Should subjects used in the model be excluded from the evaluation cohort?

removeSubjectsWithFutureDates
For buggy data with data in the future: ignore subjects with dates in the future?

saveEvaluationCohortPlpData
Should the large PLP file for the evaluation cohort be saved? To be used for debugging purposes.

modelType
The type of health outcome in the model either "acute" or"chronic".

Details

Create an object defining the parameter values.