Create the evaluation cohort

Recalibrates a PheValuator model file

recalibrateModel(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  xSpecCohortId,
  prevalenceCohortId,
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  gender = c(8507, 8532),
  race = 0,
  ethnicity = 0,
  startDate = "19000101",
  endDate = "21000101",
  cdmVersion = "5",
  modelId = "main",
  evaluationCohortId = "main",
  modelFileDirectory = NULL,
  destinationModelFileDirectory = NULL,
  evalFileDirectory = NULL,
  destinationEvalFileDirectory = NULL,
  removeSubjectsWithFutureDates = TRUE
)

Arguments

connectionDetails

connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

cdmDatabaseSchema

The name of the database schema that contains the OMOP CDM instance. Requires read permissions to this database. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.

cohortDatabaseSchema

The name of the database schema that is the location where the cohort data used to define the at risk cohort is available. Requires read permissions to this database.

cohortTable

The tablename that contains the at risk cohort. The expectation is cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

xSpecCohortId

The number of the "extremely specific (xSpec)" cohort definition id in the cohort table (for noisy positives).

prevalenceCohortId

The number of the cohort definition id to determine the disease prevalence.

mainPopulationCohortId

The number of the cohort to be used as a base population for the model. If set to 0, the entire database population will be used.

mainPopulationCohortIdStartDay

The number of days relative to the mainPopulationCohortId cohort start date to begin including visits.

mainPopulationCohortIdEndDay

The number of days relative to the mainPopulationCohortId cohort start date to end including visits.

lowerAgeLimit

The lower age for subjects in the model.

upperAgeLimit

The upper age for subjects in the model.

gender

The gender(s) to be included.

race

The race(s) to be included.

ethnicity

The ethnicity(s) to be included.

startDate

The starting date for including subjects in the model.

endDate

The ending date for including subjects in the model.

cdmVersion

The CDM version of the database.

modelId

A string used to generate the file names for this model.

evaluationCohortId

A string used to generate the file names for this evaluation cohort.

modelFileDirectory

Directory name of the RDS model file to recalibrate

destinationModelFileDirectory

Directory name to put the recalibrated model file

evalFileDirectory

Directory name of the RDS evaluation file to get cohort

destinationEvalFileDirectory

Directory name to put the cohort RDS file

removeSubjectsWithFutureDates

For buggy data with data in the future: ignore subjects with dates in the future?

Details

Recalibrates a PheValuator model file based on a the prevalence calculated by the prevalence cohort